+++ src/main/java/micdoodle8/mods/galacticraft/core/dimension/WorldProviderOrbit.java
@@ -1162,7 +1162,7 @@ public class WorldProviderOrbit extends WorldProvider implements IOrbitDimension
 	}
 
 	@Override
-	public int getThermalLevelModifier()
+	public float getThermalLevelModifier()
 	{
 		return 0;
 	}
